Add files via upload
Through this project, which is expected to be created based on the data related to buses, 
which is a mode of passenger transportation, the data required by the passengers can be obtained 
by connecting through this application through the officials who carry out the inspection work at 
the bus stop. To use this application, the officer on duty will be given the opportunity. After telling 
the information required by the passenger to the officer on duty, he will access this system, check 
the information and provide the information to the passengers. Also, it is possible for the 
passengers to get the information through the official there at any time. The official also enters 
the data for the application. The information of the bus is entered in a database and then the data 
in the database is used to enter the travel times. The relevant bus travel times, bus details, return 
times from the destination, date, route, route license number and the name of the bus should be 
entered.
Also, this has provided an opportunity to select the relevant bus from the relevant 
database while entering the information of the travel times. It also provides the ease of data entry 
for the officer. The data entered by the officer is stored in the database and the officer has been 
given the ability to check the data as per the needs of the officer. After providing the starting 
station, destination and date, the relevant information will be displayed to the official. Later, the 
official will be able to inform the passengers with that information.